Move hook to base
