+++ assets/js/base/hooks/index.js
@@ -9,6 +9,7 @@ export * from './use-collection-data';
 export * from './use-previous';
 export * from './use-shallow-equal';
 export * from './use-store-products';
+export * from './use-store-add-to-cart';
 export * from './use-store-notices';
 export * from './use-query-state';
 export * from './use-throw-error';
